Refactor: Ticketing Flow Package
