--- src/main/java/ticketaka/mtvs3_final_backend/concert/command/application/dto/ConcertRequestDTO.java
@@ -1,17 +1,0 @@
-package ticketaka.mtvs3_final_backend.concert.command.application.dto;
-
-public class ConcertRequestDTO {
-
-    public record entranceConcertDTO(
-            String concertName
-    ) {
-    }
-
-    public record enterDeliveryAddressDTO(
-            String userName,
-            String userPhoneNumber,
-            String userAddress1,
-            String userAddress2
-    ) {
-    }
-}
