--- src/main/java/ticketaka/mtvs3_final_backend/concert/command/domain/model/ConcertStatus.java
@@ -1,6 +1,0 @@
-package ticketaka.mtvs3_final_backend.concert.command.domain.model;
-
-public enum ConcertStatus {
-    // 준비 중, 예약 중, 공연 중
-    PREPARING, RESERVING, PERFORMING
-}
