--- src/main/java/ticketaka/mtvs3_final_backend/seat/command/application/dto/SeatDTO.java
@@ -1,10 +1,0 @@
-package ticketaka.mtvs3_final_backend.seat.command.application.dto;
-
-public class SeatDTO {
-
-    public record getSeatId(
-            String section,
-            String number
-    ) {
-    }
-}
