--- src/main/java/ticketaka/mtvs3_final_backend/seat/command/application/dto/SeatRequestDTO.java
@@ -1,4 +1,0 @@
-package ticketaka.mtvs3_final_backend.seat.command.application.dto;
-
-public class SeatRequestDTO {
-}
