--- src/main/java/ticketaka/mtvs3_final_backend/seat/query/repository/SeatQueryRepository.java
@@ -1,9 +1,0 @@
-package ticketaka.mtvs3_final_backend.seat.query.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-import ticketaka.mtvs3_final_backend.seat.command.domain.model.Seat;
-
-@Repository
-public interface SeatQueryRepository extends JpaRepository<Seat, Long> {
-}
