--- src/main/java/ticketaka/mtvs3_final_backend/ticket/command/application/dto/TicketResponseDTO.java
@@ -1,4 +1,0 @@
-package ticketaka.mtvs3_final_backend.ticket.command.application.dto;
-
-public class TicketResponseDTO {
-}
