--- src/main/java/ticketaka/mtvs3_final_backend/ticket/command/application/service/TicketService.java
@@ -1,16 +1,0 @@
-package ticketaka.mtvs3_final_backend.ticket.command.application.service;
-
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-import ticketaka.mtvs3_final_backend.ticket.command.domain.repository.TicketRepository;
-
-@Slf4j
-@Transactional(readOnly = true)
-@RequiredArgsConstructor
-@Service
-public class TicketService {
-
-    private final TicketRepository ticketRepository;
-}
