--- src/main/java/ticketaka/mtvs3_final_backend/ticket/command/domain/model/Ticket.java
@@ -1,18 +1,0 @@
-package ticketaka.mtvs3_final_backend.ticket.command.domain.model;
-
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import ticketaka.mtvs3_final_backend.BaseTimeEntity;
-
-@Getter
-@Entity
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Table(name = "ticket_tb")
-public class Ticket extends BaseTimeEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-}
