--- src/main/java/ticketaka/mtvs3_final_backend/ticket/command/domain/repository/TicketRepository.java
@@ -1,9 +1,0 @@
-package ticketaka.mtvs3_final_backend.ticket.command.domain.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-import ticketaka.mtvs3_final_backend.ticket.command.domain.model.Ticket;
-
-@Repository
-public interface TicketRepository extends JpaRepository<Ticket, Long> {
-}
