+++ src/main/java/ticketaka/mtvs3_final_backend/ticketing/concert/command/domain/repository/ConcertRepository.java
@@ -1,0 +1,13 @@
+package ticketaka.mtvs3_final_backend.ticketing.concert.command.domain.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+import ticketaka.mtvs3_final_backend.ticketing.concert.command.domain.model.Concert;
+
+import java.util.Optional;
+
+@Repository
+public interface ConcertRepository extends JpaRepository<Concert, Long> {
+
+    Optional<Concert> findByName(String concertName);
+}
