+++ src/main/java/ticketaka/mtvs3_final_backend/ticketing/seat/command/application/dto/SeatDTO.java
@@ -1,0 +1,10 @@
+package ticketaka.mtvs3_final_backend.ticketing.seat.command.application.dto;
+
+public class SeatDTO {
+
+    public record getSeatId(
+            String section,
+            String number
+    ) {
+    }
+}
