+++ src/main/java/ticketaka/mtvs3_final_backend/ticketing/seat/command/application/dto/SeatRequestDTO.java
@@ -1,0 +1,4 @@
+package ticketaka.mtvs3_final_backend.ticketing.seat.command.application.dto;
+
+public class SeatRequestDTO {
+}
