+++ src/main/java/ticketaka/mtvs3_final_backend/ticketing/seat/command/domain/model/SeatStatus.java
@@ -1,0 +1,6 @@
+package ticketaka.mtvs3_final_backend.ticketing.seat.command.domain.model;
+
+public enum SeatStatus {
+    // 접수 가능, 접수 불가, 예약됨
+    AVAILABLE, UNAVAILABLE, RESERVED
+}
