+++ src/main/java/ticketaka/mtvs3_final_backend/ticketing/ticket/command/application/dto/TicketResponseDTO.java
@@ -1,0 +1,4 @@
+package ticketaka.mtvs3_final_backend.ticketing.ticket.command.application.dto;
+
+public class TicketResponseDTO {
+}
